Add a dedicated test file for non-ZIP64 archive tests
